refactor: fix all import paths after component directory reorganization
Update relative imports in source files and vi.mock() calls in spec files
to reflect the new route-based folder hierarchy under components/.
