+++ frontend/src/components/radar/RadarPage.spec.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
 import RadarPage from "./RadarPage";
-import { api } from "../api";
-import type { RadarEntry, RadarPolicy, RadarResponse } from "../types";
+import { api } from "../../api";
+import type { RadarEntry, RadarPolicy, RadarResponse } from "../../types";
 
 vi.mock(
-	import("../api"),
+	import("../../api"),
 	() =>
 		({
 			api: {
@@ -17,7 +17,7 @@ vi.mock(
 );
 
 vi.mock(
-	import("../useCompanyLogo"),
+	import("../../useCompanyLogo"),
 	() => ({ useCompanyLogo: vi.fn(() => null) }) as any,
 );
