+++ frontend/src/components/shared/AppShell.spec.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { fireEvent, render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
 import AppShell from "./AppShell";
-import type { User } from "../types";
+import type { User } from "../../types";
 
 const mockNavigate = vi.fn();
